:truck: transfer all `treepp` and `debug` utils-related logic to the separate package
