+++ bitcoin-splitter/Cargo.toml
@@ -8,27 +8,9 @@ edition = "2021"
 bitcoin              = { workspace = true, features = ["rand-std"]}
 bitcoin-script       = { git = "https://github.com/BitVM/rust-bitcoin-script" }
 bitcoin-scriptexec   = { path = "../bitcoin-scriptexec" }
-bitcoin-script-stack = { git = "https://github.com/FairgateLabs/rust-bitcoin-script-stack"}
 
 # BitVM scripts
-bitcoin-window-mul = { git = "https://github.com/distributed-lab/bitcoin-window-mul.git" }
+bitcoin-utils      = { path = "../bitcoin-utils" }
 
 # General-purpose libraries
-strum          = "0.26"
-strum_macros   = "0.26"
-serde          = { version = "1.0.197", features = ["derive"] }
-serde_json     = "1.0.116"
-tokio          = { version = "1.37.0", features = ["full"] }
 indicatif      = "0.17.8" # Progress bar
-
-# Crypto libraries
-hex        = "0.4.3"
-sha2       = "0.10.8"
-num-bigint = { version = "0.4.4", features = ["rand"] }
-num-traits = "0.2.18"
-
-# Random libraries
-rand_chacha = "0.3.1"
-rand        = "0.8.5"
-ark-std     = "0.4.0"
-konst       = "0.3.9"
