+++ bitcoin-splitter/src/lib.rs
@@ -1,22 +1,8 @@
-#[allow(dead_code)]
-// Re-export what is needed to write treepp scripts
-pub mod treepp {
-    pub use crate::debug::{execute_script, run};
-    pub use bitcoin_script::{define_pushable, script};
-
-    define_pushable!();
-    pub use bitcoin::ScriptBuf as Script;
-}
-
 pub mod split;
-pub mod utils;
-
-pub(crate) mod debug;
-pub(crate) mod pseudo;
 
 #[cfg(test)]
 mod tests {
-    use super::treepp::*;
+    use bitcoin_utils::treepp::*;
 
     /// Tests that checks that environment was set up correctly by running a 2+3=5 script.
     #[test]
