+++ bitcoin-testscripts/Cargo.toml
@@ -5,22 +5,12 @@ edition = "2021"
 
 [dependencies]
 # Bitcoin Libraries
-bitcoin              = { workspace = true, features = ["rand-std"]}
-bitcoin-script       = { git = "https://github.com/BitVM/rust-bitcoin-script" }
-bitcoin-scriptexec   = { path = "../bitcoin-scriptexec" }
-bitcoin-script-stack = { git = "https://github.com/FairgateLabs/rust-bitcoin-script-stack"}
+bitcoin = { workspace = true, features = ["rand-std"]}
 
 # BitVM scripts
 bitcoin-window-mul = { git = "https://github.com/distributed-lab/bitcoin-window-mul.git" }
 bitcoin-splitter   = { path = "../bitcoin-splitter" }
-
-# General-purpose libraries
-strum          = "0.26"
-strum_macros   = "0.26"
-serde          = { version = "1.0.197", features = ["derive"] }
-serde_json     = "1.0.116"
-tokio          = { version = "1.37.0", features = ["full"] }
-indicatif      = "0.17.8" # Progress bar
+bitcoin-utils      = { path = "../bitcoin-utils" }
 
 # Crypto libraries
 hex        = "0.4.3"
@@ -31,5 +21,3 @@ num-traits = "0.2.18"
 # Random libraries
 rand_chacha = "0.3.1"
 rand        = "0.8.5"
-ark-std     = "0.4.0"
-konst       = "0.3.9"
