+++ bitcoin-testscripts/src/bitvm/hash/sha256.rs
@@ -8,8 +8,8 @@ use crate::bitvm::u32::{
     u32_std::{u32_drop, u32_fromaltstack, u32_pick, u32_push, u32_toaltstack},
     u32_xor::{u32_xor, u8_drop_xor_table, u8_push_xor_table},
 };
-use crate::pseudo::push_to_stack;
-use crate::treepp::*;
+use bitcoin_utils::pseudo::push_to_stack;
+use bitcoin_utils::treepp::*;
 
 const K: [u32; 64] = [
     0x428a2f98, 0x71374491, 0xb5c0fbcf, 0xe9b5dba5, 0x3956c25b, 0x59f111f1, 0x923f82a4, 0xab1c5ed5,
