+++ bitcoin-testscripts/src/bitvm/u32/u32_and.rs
@@ -1,7 +1,7 @@
 #![allow(dead_code)]
 
 use crate::bitvm::u32::u32_zip::u32_copy_zip;
-use crate::treepp::*;
+use bitcoin_utils::treepp::*;
 
 /// The bitwise AND of two u8 elements.
 /// Expects the u8_xor_table to be on the stack
@@ -101,8 +101,8 @@ mod tests {
     use crate::bitvm::u32::u32_and::*;
     use crate::bitvm::u32::u32_std::*;
     use crate::bitvm::u32::u32_xor::{u8_drop_xor_table, u8_push_xor_table};
-    use crate::debug::run_and_assert;
-    use crate::treepp::script;
+    use bitcoin_utils::debug::run_and_assert;
+    use bitcoin_utils::treepp::script;
     use rand::Rng;
 
     fn and(x: u32, y: u32) -> u32 {
