+++ bitcoin-testscripts/src/int_mul_karatsuba.rs
@@ -2,8 +2,8 @@
 //! for performing the multiplication of two large integers
 //! (exceeding standard Bitcoin 31-bit integers)
 
-use crate::treepp::*;
 use bitcoin_splitter::split::script::{IOPair, SplitableScript};
+use bitcoin_utils::treepp::*;
 use bitcoin_window_mul::{
     bigint::{arithmetics::u29x9::u29x9_mul_karazuba, implementation::NonNativeBigIntImpl},
     traits::integer::{NonNativeInteger, NonNativeLimbInteger},
@@ -71,8 +71,8 @@ impl SplitableScript<{ INPUT_SIZE }, { OUTPUT_SIZE }> for U261MulKaratsubaScript
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::utils::stack_to_script;
     use bitcoin_splitter::split::core::SplitType;
+    use bitcoin_utils::stack_to_script;
     use bitcoin_window_mul::traits::comparable::Comparable;
 
     #[test]
