+++ bitcoin-utils/src/comparison.rs
