+++ bitcoin-winternitz/Cargo.toml
@@ -8,6 +8,7 @@ bitcoin.workspace = true
 bitvec = { version = "1.0.1", default-features = false, features = ["alloc"] }
 rand = { version = "0.8.5", default-features = false, optional = true, features = ["min_const_gen"] }
 bitcoin-splitter.path = "../bitcoin-splitter"
+bitcoin-utils.path = "../bitcoin-utils"
 
 [features]
 default = ["rand"]
