+++ core/src/lib.rs
@@ -1,15 +1,2 @@
-pub mod disprove;
 pub mod assert;
-pub mod debug;
-pub mod utils;
-pub mod winternitz;
-
-#[allow(dead_code)]
-// Re-export what is needed to write treepp scripts
-pub mod treepp {
-    pub use crate::debug::{execute_script, run};
-    pub use bitcoin_script::{define_pushable, script};
-
-    define_pushable!();
-    pub use bitcoin::ScriptBuf as Script;
-}
+pub mod disprove;
